baud rate check - move to serial endpoint
