+++ OpenHD/ohd_telemetry/src/endpoints/SerialEndpoint.cpp
@@ -11,6 +11,7 @@
 #include <unistd.h>
 
 #include <cassert>
+#include <map>
 #include <utility>
 
 #include "openhd_util_filesystem.h"
@@ -312,3 +313,29 @@ void SerialEndpoint::stop() {
   m_console->debug("stop()-end");
 }
 
+// based on mavsdk and what linux allows setting
+// if a value is in the map, we allow the user to set it
+static std::map<int,void*> valid_uart_baudrates(){
+  std::map<int,void*> ret;
+  ret[9600]=nullptr;
+  ret[19200]=nullptr;
+  ret[38400]=nullptr;
+  ret[57600]=nullptr;
+  ret[115200]=nullptr;
+  ret[230400]=nullptr;
+  ret[460800]=nullptr;
+  ret[500000]=nullptr;
+  ret[576000]=nullptr;
+  ret[921600]=nullptr;
+  ret[1000000]=nullptr;
+  // I think it is sane to stop here, I doubt anything higher makes sense
+  return ret;
+}
+
+bool SerialEndpoint::is_valid_linux_baudrate(int baudrate) {
+  const auto supported=valid_uart_baudrates();
+  if(supported.find(baudrate)!=supported.end()){
+    return true;
+  }
+  return false;
+}
